Initial check-in
Define instrument cluster demo platform.  This layer is similar to
meta-agl-demo, except it defines a small platform for the instrument
cluster demo, based on the agl-profile-cluster-qt5 profile.  There
is a "agl-cluster-demo" feature defined to enable building the
platform.

At the moment, only qemux86-64 and Minnowboard turbot versions have
been built and tested.  Other architectures should be feasible, but the
agl-cluster-demo-receiver application needs some further development to
support non-Intel platforms, as currently it is using the vaapisink
gstreamer plug-in.

Change-Id: I217608a1ecc9678d5d55a95c65f3ab6f10d5b54f
Signed-off-by: Scott Murray <scott.murray@konsulko.com>